[DataGeneration] Refactor: поменял расположение файлов
