+++ src/main/java/project/configuration/AppConfiguration.java
@@ -4,11 +4,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-import org.springframework.context.annotation.Profile;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import project.configuration.web.WebConfig;
-import project.repository.multiplier.TestDataGenerator;
 
 import javax.sql.DataSource;
