+++ src/main/java/project/multiplier/controller/DataGenerationController.java
@@ -1,4 +1,4 @@
-package project.controller;
+package project.multiplier.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import project.repository.multiplier.TestDataGenerator;
+import project.multiplier.service.TestDataGenerator;
 
 @RestController
 @RequestMapping("/service")
